+++ src/configStores/index.ts
@@ -1,4 +1,3 @@
-export { default as InternalConfigStore, initialConfig as internalConfig } from './InternalConfigStore';
 export { initialConfig as retaxConfig } from './RetaxConfigStore';
 export { default as DomRetaxConfigStore } from './DomRetaxConfigStore';
 export { default as RequestRetaxConfigStore } from './RequestRetaxConfigStore';