+++ src/configStores/interfaces/index.ts
@@ -1,2 +1,1 @@
-export * from './internalConfig';
 export * from './retaxConfig';