+++ src/constants/internalConfig.ts
@@ -1,0 +1,2 @@
+export const COOKIE_AUTH_TOKEN_KEY: string = 'auth_token';
+export const INITIAL_STATE_KEY: string = '__INITIAL_STATE__';