+++ src/cookieProxies/__tests__/DomCookieProxy.ts
@@ -1,49 +1,45 @@
 jest.unmock('inversify');
+jest.unmock('../../constants');
+jest.unmock('../../constants/internalConfig');
 jest.unmock('../DomCookieProxy');
 jest.unmock('../CookieProxy');
 
+import { COOKIE_AUTH_TOKEN_KEY } from '../../constants';
 import DomCookieProxy from '../DomCookieProxy';
 import * as Cookie from 'js-cookie';
 
 describe('DomCookieProxy', () => {
-  const configStore = {
-    config: {
-      COOKIE_AUTH_TOKEN_KEY: 'auth_token',
-      INITIAL_STATE_KEY: '__INITIAL_STATE__',
-    },
-  };
-
   it('reads the current auth token', () => {
-    const cookieProxy = new DomCookieProxy(<any>configStore);
+    const cookieProxy = new DomCookieProxy();
 
     /* tslint:disable */
     const { authToken } = cookieProxy;
     /* tslint:enable */
 
     expect(Cookie.get).toBeCalledWith(
-      configStore.config.COOKIE_AUTH_TOKEN_KEY
+      COOKIE_AUTH_TOKEN_KEY
     );
   });
 
   it('set the auth token', () => {
-    const cookieProxy = new DomCookieProxy(<any>configStore);
+    const cookieProxy = new DomCookieProxy();
 
     cookieProxy.authToken = '1234';
 
     expect(Cookie.set).toBeCalledWith(
-      configStore.config.COOKIE_AUTH_TOKEN_KEY,
+      COOKIE_AUTH_TOKEN_KEY,
       '1234',
       { expires: 1 }
     );
   });
 
   it('removes the auth token', () => {
-    const cookieProxy = new DomCookieProxy(<any>configStore);
+    const cookieProxy = new DomCookieProxy();
 
     cookieProxy.deleteAuthToken();
 
     expect(Cookie.remove).toBeCalledWith(
-      configStore.config.COOKIE_AUTH_TOKEN_KEY
+      COOKIE_AUTH_TOKEN_KEY
     );
   });
 });