+++ src/cookieProxies/__tests__/RequestCookieProxy.ts
@@ -1,21 +1,18 @@
 jest.unmock('inversify');
+jest.unmock('../../constants');
+jest.unmock('../../constants/internalConfig');
 jest.unmock('../RequestCookieProxy');
 jest.unmock('../CookieProxy');
 
+import { COOKIE_AUTH_TOKEN_KEY } from '../../constants';
 import RequestCookieProxy from '../RequestCookieProxy';
 
 describe('RequestCookieProxy', () => {
-  const configStore = {
-    config: {
-      COOKIE_AUTH_TOKEN_KEY: 'auth_token',
-      INITIAL_STATE_KEY: '__INITIAL_STATE__',
-    },
-  };
   const context = {
     request: {
       req: {
         cookies: {
-          [configStore.config.COOKIE_AUTH_TOKEN_KEY]: '1234',
+          [COOKIE_AUTH_TOKEN_KEY]: '1234',
         },
       },
       res: {
@@ -25,34 +22,34 @@ describe('RequestCookieProxy', () => {
   };
 
   it('reads the current auth token', () => {
-    const cookieProxy = new RequestCookieProxy(<any>configStore, <any>context);
+    const cookieProxy = new RequestCookieProxy(<any>context);
 
     /* tslint:disable */
     const { authToken } = cookieProxy;
     /* tslint:enable */
 
-    expect(context.request.req.cookies[configStore.config.COOKIE_AUTH_TOKEN_KEY])
+    expect(context.request.req.cookies[COOKIE_AUTH_TOKEN_KEY])
       .toEqual('1234');
   });
 
   it('set the auth token', () => {
-    const cookieProxy = new RequestCookieProxy(<any>configStore, <any>context);
+    const cookieProxy = new RequestCookieProxy(<any>context);
 
     cookieProxy.authToken = '1234';
 
     expect(context.request.res.cookie).toBeCalledWith(
-      configStore.config.COOKIE_AUTH_TOKEN_KEY,
+      COOKIE_AUTH_TOKEN_KEY,
       '1234'
     );
   });
 
   it('removes the auth token', () => {
-    const cookieProxy = new RequestCookieProxy(<any>configStore, <any>context);
+    const cookieProxy = new RequestCookieProxy(<any>context);
 
     cookieProxy.deleteAuthToken();
 
     expect(context.request.res.cookie).toBeCalledWith(
-      configStore.config.COOKIE_AUTH_TOKEN_KEY,
+      COOKIE_AUTH_TOKEN_KEY,
       undefined
     );
   });