+++ src/inversify/modules/common.ts
@@ -1,7 +1,6 @@
 import { IKernel, IKernelModule } from 'inversify';
 
 import { RetaxProvider, ILifecycleService, ILifecycleServiceConstructor } from '../../components';
-import { IInternalConfigStore, InternalConfigStore } from '../../configStores';
 import { IReactRouterFacade, ReactRouterFacade } from '../../reactRouter';
 import { IReduxFacade, ReduxFacade } from '../../redux';
 import { IRetaxMediator, RetaxMediator } from '../../mediator';
@@ -10,7 +9,6 @@ import { IContext } from '../../context';
 import {
   MEDIATOR,
   COMPONENTS,
-  CONFIG_STORES,
   REACT_ROUTER_FACADE,
   REDUX_FACADE,
   CONTEXT,
@@ -23,8 +21,6 @@ export default function commonModule(kernel: IKernel): void {
   kernel.bind<IReduxFacade>(REDUX_FACADE).to(ReduxFacade).inSingletonScope();
 
   kernel.bind<typeof RetaxProvider>(COMPONENTS.RETAX_PROVIDER_COMPONENT).toConstructor(RetaxProvider);
-
-  kernel.bind<IInternalConfigStore>(CONFIG_STORES.INTERNAL_CONFIG_STORE).to(InternalConfigStore).inSingletonScope();
 }
 
 export function contextModuleFactory(context: IContext): IKernelModule {