+++ src/stateProxies/__tests__/DomStateProxy-test.ts
@@ -1,16 +1,12 @@
 jest.unmock('inversify');
+jest.unmock('../../constants');
+jest.unmock('../../constants/internalConfig');
 jest.unmock('../DomStateProxy');
 
+import { INITIAL_STATE_KEY } from '../../constants';
 import DomStateProxy from '../DomStateProxy';
 
 describe('DomStateProxy', () => {
-  const internalConfigStore = {
-    config: {
-      COOKIE_AUTH_TOKEN_KEY: 'auth_token',
-      INITIAL_STATE_KEY: '__INITIAL_STATE__',
-    },
-  };
-
   const retaxConfigStore = {
     config: {
        store: {
@@ -20,9 +16,9 @@ describe('DomStateProxy', () => {
   };
 
   it('reads the initial state and convert it to an immutable object', () => {
-    const proxy = new DomStateProxy(<any>retaxConfigStore, <any>internalConfigStore);
+    const proxy = new DomStateProxy(<any>retaxConfigStore);
 
-    window[internalConfigStore.config.INITIAL_STATE_KEY] = {
+    window[INITIAL_STATE_KEY] = {
       app: {
         here: true,
       },