+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#1.1.1
+* Remove `InternalConfigStore`. Its constants has been put in a `./constants` folder.
+  This is easier to import them into another project.
+
 #1.1.0-alpha
 * retax-core no longer depend on **webpack-isomorphic-tools**
 * all JSX builder return the app build (no more html wrapper)